OC-5860: PForm renders radio / INT.  Code cleanup + display changes recommended by Product Owner.
