+++ web/src/main/java/org/akaza/openclinica/web/pform/dto/Bind.java
@@ -11,71 +11,51 @@ public class Bind {
 	private String jrPreload;
 	
 	public String getNodeSet() {
-	  return nodeSet;
-  }
-	
+		return nodeSet;
+	}
 	public void setNodeSet(String nodeSet) {
-	  this.nodeSet = nodeSet;
-  }
-	
+		this.nodeSet = nodeSet;
+	}
 	public String getType() {
-	  return type;
-  }
-	
+		return type;
+	}
 	public void setType(String type) {
-	  this.type = type;
-  }
-	
+		this.type = type;
+	}
 	public String getReadOnly() {
-	  return readOnly;
-  }
-	
+		return readOnly;
+	}
 	public void setReadOnly(String readOnly) {
-	  this.readOnly = readOnly;
-  }
-	
+		this.readOnly = readOnly;
+	}
 	public String getCalculate() {
-	  return calculate;
-  }
-	
+		return calculate;
+	}
 	public void setCalculate(String calculate) {
-	  this.calculate = calculate;
-  }
-	
+		this.calculate = calculate;
+	}
 	public String getConstraint() {
-	  return constraint;
-  }
-	
+		return constraint;
+	}
 	public void setConstraint(String constraint) {
-	  this.constraint = constraint;
-  }
-	
+		this.constraint = constraint;
+	}
 	public String getConstraintMsg() {
-	  return constraintMsg;
-  }
-	
+		return constraintMsg;
+	}
 	public void setConstraintMsg(String constraintMsg) {
-	  this.constraintMsg = constraintMsg;
-  }
-	
+		this.constraintMsg = constraintMsg;
+	}
 	public String getRequired() {
-	  return required;
-  }
-	
+		return required;
+	}
 	public void setRequired(String required) {
-	  this.required = required;
-  }
-
+		this.required = required;
+	}
 	public String getJrPreload() {
 		return jrPreload;
 	}
-
 	public void setJrPreload(String jrPreload) {
 		this.jrPreload = jrPreload;
-	}
-
-	@Override
-	public String toString() {
-	  return "<bind nodeset="+nodeSet+" type="+type+" readonly= "+readOnly+" calculate=" +calculate+ " constraint=" +constraint+ " constraintMsg=" +constraintMsg+ " required=" +required+ " jrPreload" +jrPreload+ " />";
 	}
 }
