+++ web/src/main/java/org/akaza/openclinica/web/pform/dto/Body.java
@@ -8,37 +8,21 @@ public class Body {
 	private List<Group> group;
 
 	public List<Group> getGroup() {
-	  return group;
-  }
-
+		return group;
+	}
 	public void setGroup(List<Group> group) {
-	  this.group = group;
-  }
-	
+		this.group = group;
+	}
 	public String getCssClass() {
 		return cssClass;
 	}
-
 	public void setCssClass(String cssClass) {
 		this.cssClass = cssClass;
 	}
-
 	public String getAppearance() {
 		return appearance;
 	}
-
 	public void setAppearance(String appearance) {
 		this.appearance = appearance;
-	}
-	@Override
-	public String toString() {
-		String temp = "<body> ";
-		if (group != null) {
-			for (int i = 0; i < group.size(); i++) {
-				temp = temp + " " + group.get(i).toString();
-			}
-		}		
-		temp = temp + " </body>";
-	  return temp;
 	}
 }
