+++ web/src/main/java/org/akaza/openclinica/web/pform/dto/Group.java
@@ -8,44 +8,21 @@ public class Group {
 	private List<UserControl> usercontrol;
 	
 	public Label getLabel() {
-	  return label;
-  }
-	
+		return label;
+	}
 	public void setLabel(Label label) {
-	  this.label = label;
-  }
-		
+		this.label = label;
+	}	
 	public String getAppearance() {
 		return appearance;
 	}
-
 	public void setAppearance(String appearance) {
 		this.appearance = appearance;
 	}
-
 	public List<UserControl> getUsercontrol() {
-	  return usercontrol;
-  }
-
+		return usercontrol;
+	}
 	public void setUsercontrol(List<UserControl> usercontrol) {
-	  this.usercontrol = usercontrol;
-  }
-	
-	@Override
-	public String toString() {
-		String temp = "<group>";
-		if (label != null) {
-			temp = temp + " " + label.toString();
-		}
-		
-		
-		if (usercontrol != null) {
-			for (int i = 0; i < usercontrol.size(); i++) {
-				temp = temp + " " + usercontrol.get(i).toString();
-			}
-		}
-		
-		temp = temp + "</group>";
-	  return temp;
+		this.usercontrol = usercontrol;
 	}
 }
