+++ web/src/main/java/org/akaza/openclinica/web/pform/widget/SelectWidget.java
@@ -29,10 +29,10 @@ public class SelectWidget extends BaseWidget {
 	public UserControl getUserControl() {
 		Select select = new Select();
 		Label label = new Label();
-		label.setLabel(item.getItemMeta().getLeftItemText());
+		label.setLabel(item.getItemMeta().getHeader());
 		select.setLabel(label);
 		Hint hint = new Hint();
-		hint.setHint(item.getDescription());
+		hint.setHint(item.getItemMeta().getLeftItemText());
 		select.setHint(hint);
 
 		select.setRef("/" + version.getOid() + "/" + item.getOid());
